requirements.txt
langchain>=1.2.0,<2.0.0
langchain-core>=1.2.7,<2.0.0
langchain-community>=0.3.0,<1.0.0
langchain-groq>=0.2.0
langchain-huggingface>=1.2.0
langgraph>=0.2.0
streamlit>=1.41.0
pypdf>=4.2.0
faiss-cpu>=1.13.0
langchain-text-splitters>=0.2.0
python-dotenv>=1.0.0
